Reported by coconutrandom, Today (2 hours ago)
What steps will reproduce the problem?
1. create 2 WVectors
2. make 1 fixed
3. create WRigidConstraint with the vectors
4. run

What is the expected output? What do you see instead?
If one and only of the points is fixed, that same point will still move.
It works fine if both or neither of the points are fixed.

What version of the product are you using? On what operating system?
2.0 Alpha

To patch it, remove the !(not)s in the if statements like this:
near line 177:

if (p1.fixed){
   p2.curr=WVectorMath.sub(p1.curr,WVectorMath.scale(delta,diff));
   return;
}
if (p2.fixed) {
   p1.curr=WVectorMath.addVector(p2.curr,WVectorMath.scale(delta,diff))		}

Attached is the fixed file.